+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0.0
+
+- Added more generic `ArrayNestedList` class with methods `arrayNestedList`
+and `toNestedList` to convert between nested lists and
+multi-dimensional Arrays with any indices.
+
 ## 0.1.0.0
 
 - Initial release